requirements.txt
requests~=2.31.0
tiktoken~=0.6.0
PyYAML~=6.0.1
tqdm~=4.66.2
python-dotenv~=1.0.1
pytest~=8.1.1